Renamed read/write checkpoint method
