+++ datasync/chaindatafetcher/mocks/repository_mock.go
@@ -34,9 +34,9 @@ func (m *MockRepository) EXPECT() *MockRepositoryMockRecorder {
 }
 
 // GetCheckpoint mocks base method
-func (m *MockRepository) GetCheckpoint() (int64, error) {
+func (m *MockRepository) ReadCheckpoint() (int64, error) {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "GetCheckpoint")
+	ret := m.ctrl.Call(m, "ReadCheckpoint")
 	ret0, _ := ret[0].(int64)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
@@ -45,7 +45,7 @@ func (m *MockRepository) GetCheckpoint() (int64, error) {
 // GetCheckpoint indicates an expected call of GetCheckpoint
 func (mr *MockRepositoryMockRecorder) GetCheckpoint() *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetCheckpoint", reflect.TypeOf((*MockRepository)(nil).GetCheckpoint))
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ReadCheckpoint", reflect.TypeOf((*MockRepository)(nil).ReadCheckpoint))
 }
 
 // InsertTokenTransfers mocks base method
@@ -91,9 +91,9 @@ func (mr *MockRepositoryMockRecorder) InsertTransactions(arg0 interface{}) *gomo
 }
 
 // SetCheckpoint mocks base method
-func (m *MockRepository) SetCheckpoint(arg0 int64) error {
+func (m *MockRepository) WriteCheckpoint(arg0 int64) error {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "SetCheckpoint", arg0)
+	ret := m.ctrl.Call(m, "WriteCheckpoint", arg0)
 	ret0, _ := ret[0].(error)
 	return ret0
 }
@@ -101,5 +101,5 @@ func (m *MockRepository) SetCheckpoint(arg0 int64) error {
 // SetCheckpoint indicates an expected call of SetCheckpoint
 func (mr *MockRepositoryMockRecorder) SetCheckpoint(arg0 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SetCheckpoint", reflect.TypeOf((*MockRepository)(nil).SetCheckpoint), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "WriteCheckpoint", reflect.TypeOf((*MockRepository)(nil).WriteCheckpoint), arg0)
 }
